obliged to mvvm architecture for myfitview
